// client/src/types/PixData.ts
export interface PixData {
  valor?: string;
  destinatario?: string;
  chavePix?: string;
  data?: string;
  hora?: string;
  idTransacao?: string;
  banco?: string;
  tipoChave?: string;
  status?: string;
  observacoes?: string;
  error?: string;
  filename?: string;
  extractedText?: string;
  processedAt?: string;
  data_transferencia?: string;
}